+++ NEWS.md
@@ -9,6 +9,8 @@
 - By default, closable is now FALSE for `bs4Card()`
 - value is now mandatory in `bs4InfoBox()`
 - change `attachmentBlock()` title_url to titleUrl (to match with a new upcoming package...)
+- `descriptionBlock()`: number_color, number_icon, right_border and margin_bottom become
+camel case parms (numberColor, numberIcon, rightBorder and marginBottom)
 
 ## New features
 - New `bs4CardLayout()`: simplify the way to deal with `bs4Card()`!